Support for output widget and interact

+++ src/xutils.cpp
@@ -124,7 +124,7 @@ namespace nlohmann
                 double number = j.get<double>();
                 if (number == std::floor(number))
                 {
-                    return py::int_(j.get<int>());
+                    return py::int_(j.get<long>());
                 }
                 else
                 {
